Trajectory with one planet and "sun" in ARscene
Had to alter some small things in jussi's code, preserved original with comments
Added some comments on critical parts of the code
User should now be able to press screen wich pauses the game and shows the trajectory that the planet will have and then launch the planet with an as of now fixed velocity into this trajectory when the screen is released. Upon further touches of the screen the game should pause but the trajectory and launchvelocity should not be calculated or added.